Changed the datatype of the timestamp field from float to BigDecimal. flaot is not able to represent the timestamp used by Bitfinex precisely.

+++ xchange-bitfinex/src/main/java/com/xeiam/xchange/bitfinex/v1/dto/trade/BitfinexTradeResponse.java
@@ -8,7 +8,7 @@ public class BitfinexTradeResponse {
 
   private final BigDecimal price;
   private final BigDecimal amount;
-  private final float timestamp;
+  private final BigDecimal timestamp;
   private final String exchange;
   private final String type;
   private final String tradeId;
@@ -25,7 +25,7 @@ public class BitfinexTradeResponse {
    * @param tradeId
    * @param orderId
    */
-  public BitfinexTradeResponse(@JsonProperty("price") final BigDecimal price, @JsonProperty("amount") final BigDecimal amount, @JsonProperty("timestamp") final float timestamp,
+  public BitfinexTradeResponse(@JsonProperty("price") final BigDecimal price, @JsonProperty("amount") final BigDecimal amount, @JsonProperty("timestamp") final BigDecimal timestamp,
       @JsonProperty("exchange") final String exchange, @JsonProperty("type") final String type, @JsonProperty("tid") final String tradeId, @JsonProperty("order_id") final String orderId) {
 
     this.price = price;
@@ -47,7 +47,7 @@ public class BitfinexTradeResponse {
     return amount;
   }
 
-  public float getTimestamp() {
+  public BigDecimal getTimestamp() {
 
     return timestamp;
   }
